# coreservice/internal/repository/sqlc/generated/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.28.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type DailyTask struct {
	TaskDate       pgtype.Date
	ReferalsAmount pgtype.Int4
	ReferalsReward pgtype.Numeric
	WinsAmount     pgtype.Int4
	WinReward      pgtype.Numeric
}

type Leaderboard struct {
	SeasonID int64
	UserID   int32
	Win      int32
}

type Season struct {
	ID           int64
	SeasonStart  pgtype.Timestamptz
	SeasonEnd    pgtype.Timestamptz
	SeasonFund   pgtype.Int4
	SeasonStatus pgtype.Text
}

type TaskStatusTbl struct {
	TaskDate       pgtype.Date
	UserID         int32
	Win            pgtype.Int4
	WinStatus      bool
	Referals       pgtype.Int4
	ReferalsStatus bool
}

type User struct {
	ID      int32
	Login   string
	Balance pgtype.Numeric
	WinRate pgtype.Numeric
}